Downgrade libreoffice
We were using a custom version of libreoffice because of certain
requirements which are no longer valid.

In the latest libreoffice (5.1) in Impress when we choose View Toolbars
-> Text. This opens a floating window. With 5.1, clicking on that window
creates a move event, which our window manager (ratpoison) does not
allow, and that causes libreoffice to crash. For now we're just
reverting our LO so that it does not crash.

In the future, we will need to either fix LO or the window manager.

This also fixes #3