Modify libreoffice skel files to add macro that notifies file path changes
We attach this macro to Open Office "Title changed" event. Each time we
open a new document or we save it with another name it will get the new
name and send it to the desktop.

We want it because we want the desktop to display always the correct
file in the url. Doing so when the user refreshes the page will always
see the correct file.